telemetry cleanup
