+++ nex/node_up_preflight.go
@@ -14,9 +14,9 @@ func setConditionalCommands() {
 	nodeUp.Flag("config", "configuration file for the node").Default("./config.json").StringVar(&NodeOpts.ConfigFilepath)
 	nodeUp.Flag("metrics", "enable open telemetry metrics endpoint").Default("false").UnNegatableBoolVar(&NodeOpts.OtelMetrics)
 	nodeUp.Flag("metrics_port", "enable open telemetry metrics endpoint").Default("8085").IntVar(&NodeOpts.OtelMetricsPort)
-	nodeUp.Flag("otel_metrics_exporter", "OTel exporter for metrics").Default("stdout").EnumVar(&NodeOpts.OtelMetricsExporter, "stdout", "prometheus")
+	nodeUp.Flag("otel_metrics_exporter", "OTel exporter for metrics").Default("file").EnumVar(&NodeOpts.OtelMetricsExporter, "file", "prometheus")
 	nodeUp.Flag("traces", "enable open telemetry traces").Default("false").UnNegatableBoolVar(&NodeOpts.OtelTraces)
-	nodeUp.Flag("otel_traces_exporter", "OTel exporter for tracess").Default("grpc").EnumVar(&NodeOpts.OtelTracesExporter, "grpc")
+	nodeUp.Flag("otel_traces_exporter", "OTel exporter for traces").Default("file").EnumVar(&NodeOpts.OtelTracesExporter, "file", "grpc")
 
 	nodePreflight = nodes.Command("preflight", "Checks system for node requirements and installs missing")
 	nodePreflight.Flag("force", "installs missing dependencies without prompt").Default("false").BoolVar(&NodeOpts.ForceDepInstall)
@@ -25,6 +25,7 @@ func setConditionalCommands() {
 
 func RunNodeUp(ctx context.Context, logger *slog.Logger) error {
 	ctx, cancel := context.WithCancel(newContext(ctx))
+	logger.Debug("******************* config", slog.String("metrics", NodeOpts.OtelMetricsExporter), slog.String("traces", NodeOpts.OtelTracesExporter))
 	err := nexnode.CmdUp(Opts, NodeOpts, ctx, cancel, logger)
 	if err != nil {
 		return err
